Remove Debug statements and update folder structure
