+++ AoTRetaliationOfTheSurveyCorps/Assets/Characters/Player/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4cc42c99b1dc15438761ab4aee4c253
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
