+++ AoTRetaliationOfTheSurveyCorps/Assets/Characters/Player/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d35873aacddc7ba49af40127ec29d0e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
