+++ AoTRetaliationOfTheSurveyCorps/Assets/OdmGear/GrappleHooks/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d6e91f82fd5ee441aa1e564b60a8237
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
