+++ AoTRetaliationOfTheSurveyCorps/Assets/OdmGear/GrappleHooks/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1a688474f2357c441bc0758800cf8471
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
