+++ AoTRetaliationOfTheSurveyCorps/Assets/OdmGear/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c55572442029ec7499b0bef1b2fb1843
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
